+++ Model/Configuration.php
@@ -26,6 +26,9 @@ class Configuration
     // Store configuration key for Redirect to Flow Checkout
     const FLOW_REDIRECT_ENABLED = 'flowcommerce/flowconnector/redirect_enabled';
 
+    // Store configuration key for Supporting Magento Discounts Flow Checkout
+    const FLOW_SUPPORT_MAGENTO_DISCOUNTS = 'flowcommerce/flowconnector/support_magento_discounts';
+
     // Store configuration key for Flow Invoice Event
     const FLOW_INVOICE_EVENT = 'flowcommerce/flowconnector/invoice_event';
 
@@ -67,6 +70,12 @@ class Configuration
 
     // Store configuration key for daily catalog syncing
     const FLOW_DAILY_CATALOG_SYNC = 'flowcommerce/flowconnector/daily_catalog_sync';
+
+    // Store configuration key for preloading localized catalog cache
+    const FLOW_PRELOAD_LOCALIZED_CATALOG_CACHE = 'flowcommerce/flowconnector/preload_localized_catalog_cache';
+
+    // Store configuration key for overriding final prices with regular prices
+    const FLOW_REGULAR_PRICING_OVERRIDE = 'flowcommerce/flowconnector/regular_pricing_override';
 
     // Store configuration key for checkout base url
     const FLOW_CHECKOUT_BASE_URL = 'flowcommerce/flowconnector/checkout_base_url';
@@ -171,6 +180,24 @@ class Configuration
         }
         return (bool) $this->scopeConfig->getValue(
             self::FLOW_REDIRECT_ENABLED,
+            ScopeInterface::SCOPE_STORE,
+            $storeId
+        );
+    }
+
+    /**
+     * Returns true if Support Magento Discounts is enabled in the Admin Store Configuration.
+     * @param int|null $storeId
+     * @return bool
+     * @throws NoSuchEntityException
+     */
+    public function isSupportMagentoDiscounts($storeId = null)
+    {
+        if ($storeId === null) {
+            $storeId = $this->getCurrentStoreId();
+        }
+        return (bool) $this->scopeConfig->getValue(
+            self::FLOW_SUPPORT_MAGENTO_DISCOUNTS,
             ScopeInterface::SCOPE_STORE,
             $storeId
         );
@@ -508,6 +535,44 @@ class Configuration
 
         return (bool) $this->scopeConfig->getValue(
             self::FLOW_DAILY_CATALOG_SYNC,
+            ScopeInterface::SCOPE_STORE,
+            $storeId
+        );
+    }
+
+    /**
+     * Returns true if preload localized catalog cache is enabled
+     * @param int|null $storeId
+     * @return bool
+     * @throws NoSuchEntityException
+     */
+    public function isPreloadLocalizedCatalogCacheEnabled($storeId = null)
+    {
+        if ($storeId === null) {
+            $storeId = $this->getCurrentStoreId();
+        }
+
+        return (bool) $this->scopeConfig->getValue(
+            self::FLOW_PRELOAD_LOCALIZED_CATALOG_CACHE,
+            ScopeInterface::SCOPE_STORE,
+            $storeId
+        );
+    }
+
+    /**
+     * Returns true if preload localized catalog cache is enabled
+     * @param int|null $storeId
+     * @return bool
+     * @throws NoSuchEntityException
+     */
+    public function isRegularPricingOverride($storeId = null)
+    {
+        if ($storeId === null) {
+            $storeId = $this->getCurrentStoreId();
+        }
+
+        return (bool) $this->scopeConfig->getValue(
+            self::FLOW_REGULAR_PRICING_OVERRIDE,
             ScopeInterface::SCOPE_STORE,
             $storeId
         );